refactor: MailService 구조 리팩토링
